fix #2: update client with security token

+++ client/model_plan_detailed_cases.go
@@ -1,6 +1,6 @@
 package client
 
 type PlanDetailedCases struct {
-	CaseId     int32 `json:"case_id,omitempty"`
-	AssigneeId int32 `json:"assignee_id,omitempty"`
+	CaseId     int64 `json:"case_id,omitempty"`
+	AssigneeId int64 `json:"assignee_id,omitempty"`
 }
