+++ client/model_tag_value.go
@@ -2,5 +2,5 @@ package client
 
 type TagValue struct {
 	Title      string `json:"title,omitempty"`
-	InternalId int32  `json:"internal_id,omitempty"`
+	InternalId int64  `json:"internal_id,omitempty"`
 }
